Add support for 64-bit filesystem compatible struct inode.
The struct inode of syslinux is designed for 32-bit filesystem.
However, mordern filesystem has already evolved into 64-bit,
such as XFS. So we modified it according the current linux
kernel's struct inode, which would support 64-bit file size,
block number and inode number.

+++ core/include/fs.h
@@ -98,9 +98,9 @@ struct inode {
     const char *name;		/* Name, valid for generic path search only */
     int		 refcnt;
     int          mode;   /* FILE , DIR or SYMLINK */
-    uint32_t     size;
-    uint32_t	 blocks; /* How many blocks the file take */
-    uint32_t     ino;    /* Inode number */
+    uint64_t     size;
+    uint64_t	 blocks; /* How many blocks the file take */
+    uint64_t     ino;    /* Inode number */
     uint32_t     atime;  /* Access time */
     uint32_t     mtime;  /* Modify time */
     uint32_t     ctime;  /* Create time */
